Move repl to separate package
